refactor(components): refactor progress
