+++ packages/components/progress/__tests__/progress.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils'
 import { CircleClose } from '@element-plus/icons-vue'
-import Progress from '../src/index.vue'
+import Progress from '../src/progress.vue'
 describe('Progress.vue', () => {
   test('percent', () => {
     const wrapper = mount(Progress, {
