+++ packages/components/progress/index.ts
@@ -1,13 +1,7 @@
-import Progress from './src/index.vue'
+import { withInstall } from '@element-plus/utils/with-install'
+import Progress from './src/progress.vue'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils/types'
+export const ElProgress = withInstall(Progress)
+export default ElProgress
 
-Progress.install = (app: App): void => {
-  app.component(Progress.name, Progress)
-}
-
-const _Progress = Progress as SFCWithInstall<typeof Progress>
-
-export default _Progress
-export const ElProgress = _Progress
+export * from './src/progress'
